requirements.txt
addict==2.2.0
certifi==2018.10.15
chardet==3.0.4
colorama==0.4.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-cors-headers==2.4.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jsonapi==2.4.0
drf-autodocs==0.4.4
httpie==0.9.9
idna==2.7
image==1.5.27
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.0
oauthlib==2.1.0
Pillow==5.3.0
psycopg2==2.7.5
Pygments==2.2.0
python3-openid==3.1.0
pytz==2018.5
requests==2.20.0
requests-oauthlib==1.0.0
ruamel.yaml==0.15.78
six==1.11.0
uritemplate==3.0.0
urllib3==1.24
whitenoise==4.1.2